Use logger to show banner
